+++ src/main/java/pro/sky/java/course2/adaptedcoursework/employees/EmployeeController.java
@@ -4,37 +4,33 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+
+import java.util.Collection;
 import java.util.List;
 
 @RestController
 @RequestMapping("/employee")
 public class EmployeeController {
-    private final EmployeeService employeeService;
+    private final EmployeeServiceImpl employeeService;
 
-    public EmployeeController(EmployeeService employeeService) {
+    public EmployeeController(EmployeeServiceImpl employeeService) {
         this.employeeService = employeeService;
     }
 
     @GetMapping("/add")
     public Employee add(@RequestParam("firstName") String firstName, @RequestParam("lastName") String lastName) {
-        Employee employee = new Employee(firstName, lastName);
-        employeeService.addEmployee(firstName, lastName);
-        return employee;
+        return employeeService.addEmployee(firstName, lastName);
     }
     @GetMapping("/remove")
     public Employee remove(@RequestParam("firstName") String firstName, @RequestParam("lastName") String lastName) {
-        Employee employee = new Employee(firstName, lastName);
-        employeeService.removeEmployee(firstName, lastName);
-        return employee;
+        return employeeService.removeEmployee(firstName, lastName);
     }
     @GetMapping("/find")
     public Employee find(@RequestParam("firstName") String firstName, @RequestParam("lastName") String lastName) {
-        Employee employee = new Employee(firstName, lastName);
-        employeeService.findEmployee(firstName, lastName);
-        return employee;
+        return employeeService.findEmployee(firstName, lastName);
     }
     @GetMapping("/printAll")
-    public List<Employee> printAll(){
+    public Collection<Employee> printAll(){
         return employeeService.getEmployees();
     }
 }
